bugfixes en MRH, ahora la GUI se actualiza :D

+++ mi-ram-hq/include/graphic.h
@@ -12,10 +12,10 @@
 extern t_log* logger;
 
 void iniciar_gui(t_log*, char*);
-int crear_tripulantes(uint8_t, t_list*);
-int expulsar_tripulante(uint8_t);
+int crear_tripulantes(uint16_t, t_list*);
+int expulsar_tripulante(uint16_t);
 void chequear_errores(int);
-int mover_tripulante(uint8_t,t_posicion*);
+int mover_tripulante(uint16_t,t_posicion*);
 
 NIVEL* among_nivel;
